src/requirements.txt
flask
gumo-logging==0.0.1a0
